pyproject.toml
[project]
name = "SecretsProvider"
version = "2.0.2"
authors = [
  { name="Koen Greuell", email="koen.greuell@lifewatch.eu" },
]
description = "A package to handle secrets in Python"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
license-files = ["LICENCSE"]
dependencies = [
    "python-dotenv>=1.0.1",
    "getpass"
]

[project.urls]
Homepage = "https://github.com/NaaVRE/SecretsProvider.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"